backend/requirements.txt
pyjwt
cryptography
pydantic
pydantic_settings
fastapi[standard]